Factor out ::list() (#478)
