+++ config/services.yaml
@@ -17,6 +17,8 @@ services:
             $fileSourceFileStore: '@app.services.file_store_manager.file_source'
             $gitRepositoryFileStore: '@app.services.file_store_manager.git_repository'
             $runSourceFileStore: '@app.services.file_store_manager.run_source'
+            $fileSourceStorage: '@file_source.storage'
+            $gitRepositoryStorage: '@git_repository.storage'
 
     # makes classes in src/ available to be used as services
     # this creates a service per class whose id is the fully-qualified class name
