+++ src/Services/SerializableSourceLister.php
@@ -6,23 +6,25 @@ namespace App\Services;
 
 use App\Model\SourceFileCollection;
 use League\Flysystem\FilesystemException;
+use League\Flysystem\FilesystemReader;
 use League\Flysystem\PathNormalizer;
 
 class SerializableSourceLister
 {
     public function __construct(
+        private FileLister $fileLister,
         private PathNormalizer $pathNormalizer,
     ) {
     }
 
-    public function list(FileStoreInterface $sourceFileStore, string $path): SourceFileCollection
+    public function list(FilesystemReader $reader, string $path): SourceFileCollection
     {
         $path = $this->pathNormalizer->normalizePath($path);
 
         $sourceFiles = [];
 
         try {
-            $sourceFilePaths = $sourceFileStore->list($path, ['yml', 'yaml']);
+            $sourceFilePaths = $this->fileLister->list($reader, $path, ['yml', 'yaml']);
 
             foreach ($sourceFilePaths as $sourceFilePath) {
                 $sourceFiles[] = $this->pathNormalizer->normalizePath($path . '/' . $sourceFilePath);
