+++ tests/Functional/Services/FileStoreManagerTest.php
@@ -7,23 +7,11 @@ namespace App\Tests\Functional\Services;
 use App\Entity\FileSource;
 use App\Services\FileStoreInterface;
 use App\Tests\Model\UserId;
-use App\Tests\Services\FileStoreFixtureCreator;
 use League\Flysystem\FilesystemOperator;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 
 class FileStoreManagerTest extends WebTestCase
 {
-    private FileStoreFixtureCreator $fixtureCreator;
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $fixtureCreator = self::getContainer()->get(FileStoreFixtureCreator::class);
-        \assert($fixtureCreator instanceof FileStoreFixtureCreator);
-        $this->fixtureCreator = $fixtureCreator;
-    }
-
     /**
      * @dataProvider removeSuccessDataProvider
      */
@@ -62,99 +50,6 @@ class FileStoreManagerTest extends WebTestCase
             'run source store' => [
                 'storageServiceId' => 'run_source.storage',
                 'storeServiceId' => 'app.services.file_store_manager.run_source',
-            ],
-        ];
-    }
-
-    /**
-     * @dataProvider listDataProvider
-     *
-     * @param string[] $extensions
-     * @param string[] $expectedRelativePathNames
-     */
-    public function testListSuccess(
-        string $fixtureSet,
-        string $relativePath,
-        array $extensions,
-        array $expectedRelativePathNames
-    ): void {
-        $storage = self::getContainer()->get('file_source.storage');
-        self::assertInstanceOf(FilesystemOperator::class, $storage);
-
-        $store = self::getContainer()->get('app.services.file_store_manager.file_source');
-        self::assertInstanceOf(FileStoreInterface::class, $store);
-
-        $store->remove($relativePath);
-
-        $this->fixtureCreator->copySetTo('Source/' . $fixtureSet, $storage, $relativePath);
-
-        $files = $store->list($relativePath, $extensions);
-
-        self::assertCount(count($expectedRelativePathNames), $files);
-        self::assertSame($expectedRelativePathNames, $files);
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function listDataProvider(): array
-    {
-        $basePath = (string) new FileSource(UserId::create(), '');
-
-        return [
-            'source: txt, no extensions' => [
-                'fixtureSet' => 'txt',
-                'relativePath' => $basePath,
-                'extensions' => [],
-                'expectedRelativePathNames' => [
-                    'directory/file3.txt',
-                    'file1.txt',
-                    'file2.txt',
-                ],
-            ],
-            'source: txt, extensions=[txt]' => [
-                'fixtureSet' => 'txt',
-                'relativePath' => $basePath,
-                'extensions' => ['txt'],
-                'expected' => [
-                    'directory/file3.txt',
-                    'file1.txt',
-                    'file2.txt',
-                ],
-            ],
-            'source: txt, extensions=[yml, yaml]' => [
-                'fixtureSet' => 'txt',
-                'relativePath' => $basePath,
-                'extensions' => ['yml', 'yaml'],
-                'expected' => [],
-            ],
-            'source: yml_yaml, no extensions' => [
-                'fixtureSet' => 'yml_yaml_valid',
-                'relativePath' => $basePath,
-                'extensions' => [],
-                'expected' => [
-                    'directory/file3.yml',
-                    'file1.yaml',
-                    'file2.yml',
-                ],
-            ],
-            'source: yml_yaml, extensions=[yml]' => [
-                'fixtureSet' => 'yml_yaml_valid',
-                'relativePath' => $basePath,
-                'extensions' => ['yml'],
-                'expected' => [
-                    'directory/file3.yml',
-                    'file2.yml',
-                ],
-            ],
-            'source: mixed, extensions=[yaml]' => [
-                'fixtureSet' => 'mixed',
-                'relativePath' => $basePath,
-                'extensions' => ['yaml'],
-                'expected' => [
-                    'directory/file3.yaml',
-                    'file1.yaml',
-                ],
             ],
         ];
     }
